refactor: renamed `handlers` to `routes`
