+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod handlers;
+pub mod routes;
 pub mod models;
 pub mod utils;
 pub mod config;
