+++ src/routes/category.rs
